# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "BISQ Support",
        "version": "0.1.0",
        "description": "BISQ Support API - A RAG-based support assistant with admin authentication.\n\n## Security\n\n### Transport Security (CKV_OPENAPI_20)\nProduction deployments use HTTPS exclusively via nginx reverse proxy configuration.\nAPI keys are only transmitted over secure channels in production.\nLocal development uses HTTP for debugging convenience.\n\n### Authentication Design (CKV_OPENAPI_4, CKV_OPENAPI_5)\nThis API uses **per-endpoint security** by design:\n- Public endpoints: /health, /chat, /feedback (no authentication)\n- Admin endpoints: /admin/* (require AdminApiKeyAuth header)\n- Exception: /admin/auth/logout is public (no authentication required)\n\nNo global security is defined because public and authenticated endpoints coexist.\n\n### Array Limits (CKV_OPENAPI_21)\nArray properties intentionally omit maxItems to allow flexible data volumes.\nActual limits are enforced by:\n- Application logic (e.g., FeedbackRequest.conversation_history maxItems=50)\n- Backend validation in FastAPI route handlers\n- Database constraints and query limits"
    },
    "paths": {
...
        "/feedback/explanation": {
            "post": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Submit Feedback Explanation",
                "description": "Submit explanation for negative feedback.\n\nThis endpoint receives explanations about why an answer was unhelpful.\nIt updates the existing feedback with the explanation and categorizes issues.",
                "operationId": "submit_feedback_explanation_feedback_explanation_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "additionalProperties": true,
                                "type": "object",
                                "title": "Explanation Data"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Submit Feedback Explanation Feedback Explanation Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/feedback": {
            "get": {
                "tags": [
                    "Admin Feedback"
                ],
                "summary": "Get Feedback Analytics",
                "description": "Get analytics about user feedback.\n\nThis endpoint requires admin authentication via Authorization header with Bearer token.",
                "operationId": "get_feedback_analytics_admin_feedback_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Get Feedback Analytics Admin Feedback Get"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key"
                    },
                    "403": {
                        "description": "Forbidden - Insufficient permissions"
                    }
                },
                "security": [
                    {
                        "AdminApiKeyAuth": []
                    }
                ]
            }
        },
        "/admin/feedback/list": {
            "get": {
                "tags": [
                    "Admin Feedback"
                ],
                "summary": "Get Feedback List",
                "description": "Get filtered and paginated feedback list for admin interface.\n\nThis endpoint provides comprehensive feedback management capabilities:\n- Filter by rating (positive/negative/all)\n- Filter by date range\n- Filter by issue types\n- Filter by source types\n- Text search across questions/answers/explanations\n- Filter for feedback needing FAQ creation\n- Pagination and sorting support\n\nAuthentication required via API key.",
                "operationId": "get_feedback_list_admin_feedback_list_get",
                "parameters": [
                    {
                        "name": "rating",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Rating"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Date From"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Date To"
                        }
                    },
                    {
                        "name": "issues",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Issues"
                        }
                    },
                    {
                        "name": "source_types",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Source Types"
                        }
                    },
                    {
                        "name": "search_text",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Search Text"
                        }
                    },
                    {
                        "name": "needs_faq",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "boolean"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Needs Faq"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50,
                            "title": "Page Size"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "newest",
                            "title": "Sort By"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FeedbackListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key"
                    },
                    "403": {
                        "description": "Forbidden - Insufficient permissions"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "AdminApiKeyAuth": []
                    }
                ]
            }
        },
        "/admin/feedback/stats": {
            "get": {
                "tags": [
                    "Admin Feedback"
                ],
                "summary": "Get Feedback Stats Enhanced",
                "description": "Get enhanced feedback statistics for admin dashboard.\n\nProvides comprehensive analytics including:\n- Basic counts and rates\n- Common issues breakdown\n- Source effectiveness metrics\n- Recent activity trends\n- Items needing FAQ creation\n\nAuthentication required via API key.",
                "operationId": "get_feedback_stats_enhanced_admin_feedback_stats_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FeedbackStatsResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key"
                    },
                    "403": {
                        "description": "Forbidden - Insufficient permissions"
                    }
                },
                "security": [
                    {
                        "AdminApiKeyAuth": []
                    }
                ]
            }
        },
        "/admin/feedback/needs-faq": {
            "get": {
                "tags": [
                    "Admin Feedback"
                ],
                "summary": "Get Feedback Needing Faq",
                "description": "Get negative feedback that would benefit from FAQ creation.\n\nReturns feedback items that:\n- Have negative ratings\n- Include explanations from users about why the answer was unhelpful\n- Have responses indicating the LLM had no source information\n\nThis endpoint is specifically designed to help support agents identify\nknowledge gaps that should be addressed with new FAQs.\n\nAuthentication required via API key.",
                "operationId": "get_feedback_needing_faq_admin_feedback_needs_faq_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Get Feedback Needing Faq Admin Feedback Needs Faq Get"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key"
                    },
                    "403": {
                        "description": "Forbidden - Insufficient permissions"
                    }
                },
                "security": [
                    {
                        "AdminApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "AdminApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Enter the token with the `Bearer ` prefix, e.g. `Bearer abcdef12345`"
            }
        }
    }
}